Fixed y-flipped rendering when using a frame buffer
If the texture frame buffer (backbuffer) is rendered first,
rendering to frame buffer 0 (the screen) next will result in
a flipped image (on the y axis).

Maybe that's because rendering into a texture is (by default?)
y-flipped because a textures origin (0/0) is always in the
bottom left. And maybe the transformation matrix does not get
reset on glBindFramebuffer( GL_FRAMEBUFFER, 0 ).

Anyway, rendering into frame buffer 0 first does fix the
problem.