Wrap time variable
Time holds the number of milliseconds since the creation of the
OpenGL surface. This number can get quickly very big what pretty
soon will result in rounding errors because a mediump float is
not very long.

When that happens, results from calculations that rest on the
uniformity of the time variable will jump. The animation stutters.

To avoid this, the time variable wraps early now.