Use a Toast to display GLSL error message
Probably far better than using a hidden layout because
it won't trigger a re-layout. What makes the cursor in
the editor jump.

The pre Material Design version used a pop up window
because I thought it would be bad if the error message
vanishes by itself. On the other hand, the user can
recompile at any time so that's probably not a big
problem.