Moved settings from drawer into menu
It's probably a bit confusing having settings in the
drawer at the left (start) and the rest of the options
in the overflow menu at the right (end).

Then, when the keyboard is visible, the list of shaders
becomes quite small on small screens or landscape with
additional menu items in the drawer.

After trying the textures menu item in the drawer, it
became quite clear, that the drawer should only contain
the list of shaders and nothing else.