Removed infer rule
Unfortunately, Infer isn't quite there yet.

It's a pain to install (compliation breaks with 0.4.0) and
(at least 0.2.0) can't see if you're closing a cursor in a
method and not in the current context:

	Cursor cursor = ...

	if( closeIfEmpty( cursor ) )
		return;

You can close the cursor in closeIfEmpty() but that won't
keep Infer from complaining. There are other situations
like when you're putting a cursor into an adapter that
trigger the same kind of warning altough you're calling
adapter.changeCursor( null ); to close it.

Looking for other static analysis tools.