Transparent background color in error highlighting
Because EditText draws the cursor before the background of
Spannables, an opaque background color will make the cursor
disappear. There's an issue about that:

https://code.google.com/p/android/issues/detail?id=172001

For the time being, the simplest solution is to make the
background color transparent. It may not be as eye-catching
as before but it's probably still good enough to clearly see
where the error is.