Add READ_EXTERNAL_STORAGE permission for API < 19
Required to import textures on API < 19 from gallery apps
that do not use grantUriPermission(). Later API's will not
see that permission.

Hard decision. I don't really want to add that permission
but people (on API's < 19, especially Samsung devices)
keep running into this limitation and I just don't see a
better way to fix this.

At the time of writing, about 25% of installs will be
affected by this and will need to manually confirm the
update containing this change.