Make static members of application context final
Declaring the static members of the application context as
final is syntactically better because those members will
never change. The only reason they weren't final was, that
I preferred to make the objects only constructable with a
context object. Switched to late initialization because
it's probably more important to have the static members
to be final.