Set current shader as wallpaper shader
Change the current wallpaper shader from inside the app.
If the current shader is the wallpaper shader, the menu
item will say "Update wallpaper". Otherwise it will read
"Set as wallpaper".

Note that this only works if Shader Editor is already set
as live wallpaper. Android doesn't allow apps to set a
live wallpaper directly (so we'd need to fire an intent
for the wallpaper chooser; but that's probably too much).