Restored ScrollView around EditText
Yes, TextView can scroll already, but it's not possible to set a
padding on it's contents only. Instead, contents will be cut off
at the specified padding but before the visible border leaving
an ugly gap between text and border.

The only way to avoid this seems to be putting the TextView inside
a ScrollView.