Transfer delayed results from background tasks
Into the currently active UI object. When a device is rotated
while an AsyncTask is running, the result of that AsyncTask
should be transferred to the new Activity/Fragment.

The simplest way to do this is to use a static reference to
a private listener that is initialized in onCreate/View().
Since AsyncTask.onPostExectue() and onCreate()/onCreateView()
are both running on the main thread, they can never overlap
and the result is always realized in the latest active object.