Don't transfer shader query over config changes
Don't keep a static (hard) reference to the latest Activity object
because that would keep the GC from recycling it long after its
use. Of course, it'd be possible to use a WeakReference and avoid
leaking that Activity, but there's really no point in worrying
about having a new configuration querying shaders again.