Don't check findViewById() return values for null
Given we test every change, findViewById() can only return null
in development but not in production. Once the wiring is correct,
findViewById() cannot fail.

Even worse, if findViewById() returns null because we make a
mistake while developing, there won't be a telling exception if
we check for null and just finish the activity.