Use ACCELEROMETER if GRAVITY/LA isn't available
TYPE_GRAVITY and TYPE_LINEAR_ACCELERATION seem not be available
on all devices. Got a beta feedback from a user on a Samsung Galaxy
J2 saying that gravity and linear uniforms no longer are working
after changing to TYPE_GRAVITY and TYPE_LINEAR_ACCELERATION.

So if the app can't register one of these (software) sensors,
it'll fallback to use TYPE_ACCELEROMETER and calculate linear
and gravity from that (as before).