Release camera when opening is cancelled
Since the camera is opened in a background thread, it may
happen that the Activity enters onPause() before or while
the opening is in process. For example, this may happen
when there are a couple of quick orientation changes.

So make sure every opened camera object is properly closed.

Also better use a shortcut for the camera member variable
to avoid confusion with function parameters.