Suppress static leak lint warning for AsyncTasks
See, those AsyncTasks are *meant* to keep their parent instance
around while they're running. There's nothing wrong with keeping
the instance for a short while.

If you think different, consider if it's really worth the extra
complexity of introducing a weak reference or a bus system or
something similar. Would that make things better or just fat?
Would that "better" even be noticeable?