Add a preference to auto save changes
The editor doesn't auto-save by default because there's
no undo function. So, if you remove some code by accident,
you can always go back. For me, it's easier when saving
is deliberate when there's no undo.

Now, not everybody is me and I keep getting request to
add auto-save so here it is.