Fix undo/redo to ignore initial change
Because the editor (usually) doesn't start blank but with some source
in it. We don't want to undo before that inital change. Otherwise it
would be possible to undo before the source was loaded and end up with
a blank editor.

This requires some very specific modifications to the original
implemenation (fi.iki.asb.android.logo.TextViewUndoRedo) because of
which I moved this class into the main namespace and matched its style
with the rest of the source.

Also, the undo/redo history can now be outside of the UndoRedo class.
I did this to make it static and to not put extra load on configuration
(read orientation) changes as the original implemenation would have
done (because it saves/restores the history with shared preferences).

A downside of this is that the history is lost as soon as the app is
killed now. I still prefer to have fast configuration changes.