Automatically convert ShaderToy sources
This has been requested several times, although I see relatively
little benefit in it.

ShaderToy is hardly usable on a mobile phone. For example, I can't
even copy the source code of a shader because the JavaScript won't
let me. Also, this app only supports a small subset of ShaderToy's
uniforms. So most of the shaders can't be converted anyway.

Let's see if this really has any value for anybody.