Fix support for RTL layouts
Make sure this app is never layed out from right-to-left because
(shader) code is always layed out left-to-right. Also, there's no
translation for a RTL language at the moment, so this kind of
layout isn't helpful.

If `supportsRtl` is set to `true`, white spaces are right-aligned
and the caret positioning is off.

For details see this issue:
https://github.com/markusfisch/ShaderEditor/issues/111

Thanks to @Mehran7kl for fixing this bug!