Clean up Preferences
Line up SHOW_LINE_NUMBERS with the other names
and move the static parse functions back at the end
of the class. Why were these moved?

I'm following the original Java Language Specification that
recommends class and member modifiers, when present,
to appear in the order public, protected, private. I think this
makes sense, at least for Java, to order the put the most
interesting things to the reader at the top of the file.