Improve error line highlighting performance
Use Layout.getLineStart()/.getLineEnd() instead of manually counting
lines to highlight the error line.

As suggested by Anton Pieper in #155.

But instead of inserting a `#line 0` preprocessor to account for the
silently added preprocessor directives, this does now keep track
of how many lines were added and substracts them from the error
line to highlight the correct line in the editor.

I chose this solution because the `#line 0` preprocessor didn't work
on some of my test devices.