Distinguish between modification and edit pausing
The `OnTextChangedListener` is split into `OnTextModifiedListener` and `OnEditPausedListener` to differentiate between any modification to the text and the end of an editing session when the highlighting is updated.

This allows the `ShaderManager` to correctly track if the shader source has been modified, even if the editor fragment itself is not currently displaying unsaved changes.